Bugfix: modified type of returned value in 2 cases in Cpp_helpers.h.

+++ include/Cpp_helpers.h
@@ -206,7 +206,7 @@ inline unsigned long lossless_static_cast(int16_t x)
 {
     if (x < 0)
         throw make_error<std::runtime_error>("Can not convert: value ", x, " is out of range");
-    return static_cast<size_t>(x);
+    return static_cast<unsigned long>(x);
 }
 
 template<>
@@ -226,7 +226,7 @@ inline unsigned long lossless_static_cast(int x)
 {
     if (x < 0)
         throw make_error<std::runtime_error>("Can not convert: value ", x, " is out of range");
-    return static_cast<size_t>(x);
+    return static_cast<unsigned long>(x);
 }
 
 template<>
